Initial commit
ircservserv (aka "iss") allows for managing the state of an IRC channel,
especially various ACLs through declarative TOML files.

When asked to sync a channel, ircservserv will check to make sure all
the properties it knows about are exactly as the TOML file specifies,
otherwise it'll adjust as necessary.

ircservserv will expect to have +f access in channels it manages. It is
designed for usage with Libera Chat but hopefully doesn't hardcode any
network-specific behavior.

Change-Id: Ic7328844139dd23f7599fa1d0e3249c237a37ee6