Refactor permissions model
Any founder (+F) can trigger a !issync. The user must currently be a
founder per ChanServ, they cannot be in the list that is going to be
added as a founder.

Then there are two lists of users in the config.toml.

"owners" are the bot operators, who run the Toolforge tool. Owners can
trigger !issync in any channel.

"trusted" are trusted users who can merge in the config repo and and can
run !isspull.

Both lists can only be updated by a bot restart, so it's not possible to
escalate permissions without an owner approving it.

In theory a trusted user could attempt to hijack a channel and tricking
a founder into syncing, but there's a reason they're trusted.

Change-Id: I15c3d21179078cd030055a0933adf28413f9e6c5