Unset and set chanserv flags at the same time
Merge operations based on the user, to avoid removing modes just to
grant them again. This also fixes a small race condition in which
a flag removal could happen after granting, ending up with the wrong
flags.

Note that the output to the channel will still have the unmerged flags.

Change-Id: I1295c54d1c6512883fcd8a8b8cb3c17be93eac0d